// src/pages/upload/types.ts
export type ExcelRowType = {
  consumoAnual: number
  alcance: number
  fuenteDeConsumo: string
  subfuenteDeConsumo: string
  area: string
  unidades: string
}

export type ExcelType = {
  ' Consumo Anual ': number
  Alcance: number
  Area: string
  'Fuente de Consumo': string
  'Subfuente de Consumo': string
  Unidades: string
}